SWTBot modification: 1. modified copy method for items for genetate a right item name 2. improve creation method for items about could not find item node 3. modified after class for should not delete the system routine and sql templates
git-svn-id: http://talendforge.org/svn/tis_shared/trunk@84339 f6f1c999-d317-4740-80b0-e6d1abc6f99e

+++ org.talend.tos.swtbot.test/src/org/talend/swtbot/TalendSwtBotForTos.java
@@ -221,6 +221,8 @@ public class TalendSwtBotForTos {
                     }
                     if (ivos != null) {
                         for (final IRepositoryViewObject ivo : ivos) {
+                            if (ivo.getPath().contains("system"))
+                                continue;
                             try {
                                 factory.deleteObjectPhysical(ivo);
                             } catch (PersistenceException e) {
